fixup! gnu: Handle basic file types for 32bit with _FILE_OFFSET_BITS=64

+++ src/unix/linux_like/linux/gnu/b32/mod.rs
@@ -32,7 +32,7 @@ cfg_if! {
     } else if #[cfg(gnu_file_offset_bits64)] {
         pub type time_t = i32;
         pub type suseconds_t = i32;
-        pub type ino_t = u64;
+        pub type ino_t = u32;
         pub type off_t = i64;
         pub type blkcnt_t = i64;
         pub type fsblkcnt_t = u64;
